feat booking and commit

+++ src/main/java/ru/practicum/shareit/item/storage/ItemStorage.java
@@ -7,11 +7,11 @@ import java.util.List;
 public interface ItemStorage {
     Item createItem(Item item);
 
-    Item getItem(int itemId);
+    Item getItem(Long itemId);
 
     Item updateItem(Item item);
 
-    List<Item> getAllUserItems(int userId);
+    List<Item> getAllUserItems(Long userId);
 
     List<Item> getAllItems();
